+++ imp_llava/model/__init__.py
@@ -2,11 +2,11 @@ try:
     from .language_model.llava_llama import LlavaLlamaForCausalLM, LlavaConfig
 except:
     pass
-from .language_model.imp import ImpForCausalLM, ImpConfig
+
 try:
-    from .language_model.imp_2b import ImpQwen2ForCausalLM, ImpQwen2Config
-    from .language_model.imp_4b import ImpPhi3Config, ImpPhi3ForCausalLM
-    from .language_model.imp_newphi import Imp1_5Config, Imp1_5ForCausalLM
+    from .language_model.imp_qwen1_5 import ImpQwen2ForCausalLM, ImpQwen2Config
+    from .language_model.imp_phi3 import ImpPhi3Config, ImpPhi3ForCausalLM
+    from .language_model.imp import ImpConfig, ImpForCausalLM
 except:
     pass
