+++ imp_llava/model/builder.py
@@ -57,11 +57,7 @@ def load_pretrained_model(model_path, model_base, model_name, load_8bit=False, l
             logger.info(f'Load model name trained with LoRA, model base: {model_base}')
             assert 'imp' in model_name.lower(), 'The model name must contain `imp`'
             tokenizer = AutoTokenizer.from_pretrained(model_base, use_fast=False, trust_remote_code=True)
-            if 'v1-3b' in model_name.lower() or  'phi2' in model_name.lower():
-                if 'v1.5-3b' in model_name.lower():
-                    lora_cfg_pretrained = Imp1_5Config.from_pretrained(model_path)
-                    model = Imp1_5ForCausalLM.from_pretrained(model_base, config=lora_cfg_pretrained, **kwargs)
-                else:
+            if 'phi-2' in model_name.lower() or  'phi2' in model_name.lower():
                     lora_cfg_pretrained = ImpConfig.from_pretrained(model_path)
                     model = ImpForCausalLM.from_pretrained(model_base, config=lora_cfg_pretrained, **kwargs)
             elif 'qwen1.5' in model_name.lower():
@@ -100,28 +96,21 @@ def load_pretrained_model(model_path, model_base, model_name, load_8bit=False, l
             logger.info('Model is loaded...')
         elif model_base is not None:
             logger.info('Load mm projector only model...')
-            if 'phi2' in model_name.lower() or 'imp-v1-3b' in model_name.lower():
+            if 'phi2' in model_name.lower() or 'phi-2' in model_name.lower():
                 logger.info(f'model_base:, {model_base}')
                 config = ImpConfig.from_pretrained(model_path, trust_remote_code=True)
                 model = ImpForCausalLM.from_pretrained(model_base, **kwargs)
                 model.transformer.vision_tower = build_vision_tower(config)
                 model.transformer.mm_projector = build_vision_projector(config)
                 tokenizer = AutoTokenizer.from_pretrained(model_base)
-            elif '2b' in model_name.lower():
+            elif 'qwen1.5' in model_name.lower():
                 logger.info(f'model_base:, {model_base}')
                 config = ImpQwen2Config.from_pretrained(model_path, trust_remote_code=True)
                 model = ImpQwen2ForCausalLM.from_pretrained(model_base, **kwargs)
                 model.transformer.vision_tower = build_vision_tower(config)
                 model.transformer.mm_projector = build_vision_projector(config)
                 tokenizer = AutoTokenizer.from_pretrained(model_base)
-            elif 'new' in model_name.lower():
-                logger.info(f'model_base:, {model_base}')
-                config = Imp1_5Config.from_pretrained(model_path, trust_remote_code=True)
-                model = Imp1_5ForCausalLM.from_pretrained(model_base, **kwargs)
-                model.transformer.vision_tower = build_vision_tower(config)
-                model.transformer.mm_projector = build_vision_projector(config)
-                tokenizer = AutoTokenizer.from_pretrained(model_base)
-            elif '4b' in model_name.lower():
+            elif 'phi3' in model_name.lower():
                 logger.info(f'model_base:, {model_base}')
                 config = ImpPhi3Config.from_pretrained(model_path, trust_remote_code=True)
                 model = ImpPhi3ForCausalLM.from_pretrained(model_base, **kwargs)
@@ -143,19 +132,10 @@ def load_pretrained_model(model_path, model_base, model_name, load_8bit=False, l
             logger.info(f'load fully fine-tuned model or HF Hub model: {model_path}')
             #hg version
             
-            if 'v1-3b' in model_name.lower() or 'phi2' in model_name.lower() or 'phi-2' in model_name.lower():
-                if 'v1-3b' in model_name.lower():
-                    tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
-                    model = ImpForCausalLM.from_pretrained(model_path, **kwargs)
-                    logger.info('Model is loaded...')
-                elif 'v1.5-3b' in model_name.lower():
-                    tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
-                    model = Imp1_5ForCausalLM.from_pretrained(model_path, **kwargs)
-                    logger.info('Model is loaded...')
-                else :
-                    tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
-                    model = ImpForCausalLM.from_pretrained(model_path, **kwargs)
-                    logger.info('Model is loaded...')
+            if 'phi2' in model_name.lower() or 'phi-2' in model_name.lower():
+                tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
+                model = ImpForCausalLM.from_pretrained(model_path, **kwargs)
+                logger.info('Model is loaded...')
             elif 'qwen1.5' in model_name.lower():
                 tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
                 model = ImpQwen2ForCausalLM.from_pretrained(model_path, **kwargs)
