+++ imp_llava/model/language_model/qwen2/modeling_qwen2.py
@@ -1051,24 +1051,8 @@ class Qwen2Model(Qwen2PreTrainedModel):
                 all_hidden_states += (hidden_states,)
 
             if self.gradient_checkpointing and self.training:
-                # layer_outputs = self._gradient_checkpointing_func(
-                #     decoder_layer.__call__,
-                #     hidden_states,
-                #     attention_mask,
-                #     position_ids,
-                #     past_key_values,
-                #     output_attentions,
-                #     use_cache,
-                # )
-                def create_custom_forward(module):
-                    def custom_forward(*inputs):
-                        # None for past_key_value
-                        return module(*inputs)
-
-                    return custom_forward
-
-                layer_outputs = torch.utils.checkpoint.checkpoint(
-                    create_custom_forward(decoder_layer),
+                layer_outputs = self._gradient_checkpointing_func(
+                    decoder_layer.__call__,
                     hidden_states,
                     attention_mask,
                     position_ids,
@@ -1076,6 +1060,22 @@ class Qwen2Model(Qwen2PreTrainedModel):
                     output_attentions,
                     use_cache,
                 )
+                # def create_custom_forward(module):
+                #     def custom_forward(*inputs):
+                #         # None for past_key_value
+                #         return module(*inputs)
+
+                #     return custom_forward
+
+                # layer_outputs = torch.utils.checkpoint.checkpoint(
+                #     create_custom_forward(decoder_layer),
+                #     hidden_states,
+                #     attention_mask,
+                #     position_ids,
+                #     past_key_values,
+                #     output_attentions,
+                #     use_cache,
+                # )
             else:
                 layer_outputs = decoder_layer(
                     hidden_states,
