+++ scripts/finetune_lora_custom_oy.sh
@@ -41,6 +41,7 @@ deepspeed --include localhost:0,1,2,3 --master_port 29160 imp_llava/train/train_
     --per_device_train_batch_size 4 \
     --per_device_eval_batch_size 4 \
     --gradient_accumulation_steps 4 \
+    --gradient_checkpointing True\
     --evaluation_strategy "no" \
     --save_strategy "steps" \
     --save_steps 50000 \
